fully implement create loan screen

+++ app/src/main/java/com/ithirteeng/secondpatternsclientproject/domain/loans/model/loan/ApplicationInfo.kt
@@ -1,9 +1,9 @@
 package com.ithirteeng.secondpatternsclientproject.domain.loans.model.loan
 
 data class ApplicationInfo(
-    val clientId: Int,
-    val loanProgramId: Int,
+    val clientId: Long,
+    val loanProgramId: Long,
     val borrowedAmount: Double,
-    val loanTerm: Int,
+    val loanTerm: Long,
     val clientAccountNumber: String?,
 )
