+++ app/src/main/java/com/ithirteeng/secondpatternsclientproject/domain/loans/model/loan/LoanApplicationResponse.kt
@@ -3,8 +3,8 @@ package com.ithirteeng.secondpatternsclientproject.domain.loans.model.loan
 import com.ithirteeng.secondpatternsclientproject.domain.loans.model.program.LoanProgramInfoResponse
 
 data class LoanApplicationResponse(
-    val id: Int,
-    val clientId: Int,
+    val id: Long,
+    val clientId: Long,
     val loanProgram: LoanProgramInfoResponse,
     val borrowedAmount: Double,
     val loanTerm: Long,
