+++ app/src/main/java/com/ithirteeng/secondpatternsclientproject/features/client/myloans/loaninfo/ui/LoanInfoScreen.kt
@@ -30,7 +30,7 @@ import org.koin.androidx.compose.koinViewModel
 
 @Composable
 fun LoanInfoScreen(
-    loanId: Int,
+    loanId: Long,
     viewModel: LoanInfoViewModel = koinViewModel(),
 ) {
     val context = LocalContext.current
