+++ app/src/main/java/com/ithirteeng/secondpatternsclientproject/features/client/myloans/programinfo/presentation/model/ProgramInfoEvent.kt
@@ -5,7 +5,7 @@ import com.ithirteeng.secondpatternsclientproject.domain.loans.model.program.Loa
 
 sealed interface ProgramInfoEvent : BaseEvent {
 
-    data class Init(val programId: Int) : ProgramInfoEvent
+    data class Init(val programId: Long) : ProgramInfoEvent
 
     data class DataLoaded(val program: LoanProgramResponse) : ProgramInfoEvent
